fix nncf task
